third_party/WebKit/LayoutTests/batterystatus/basic-all-types-of-events-expected.txt

CONSOLE MESSAGE: line 17: Uncaught TypeError: Cannot call method 'addEventListener' of undefined
Tests the basic operation of all BatteryStatus events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
